options-overhaul: split up the monolithic options header file
